Creation of the terminal client module
The terminal client is the main client that will be used to play the
game, rather then being on the website, you can use the terminal client
to play the game.  The terminal client will initially be very basic,
just text for now, later on we'll add ASCII art to monsters, weapons
etc.

The terminal client is the option to use when you want to actively play
the game, but we recommend also creating automated ways of using your
character and playing the game without one for grinding purposes.