Adding get maps api and starting terminal client work
The goal of this API client will be to download the items at the start
of the game, this will include monsters and maps to make it easier to
move and view what's available to the user.

I've also started the termainal client and began work on this.  I'm
pushing changes early as development is slowing down.